change in route and controller for design

+++ app/Http/Controllers/Admin/MyBrandSliderController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Brand;
+use App\BrandDesign;
 use App\Http\Controllers\Controller;
 use App\SliderSection;
 use Illuminate\Database\Eloquent\Model;
@@ -22,19 +23,27 @@ class MyBrandSliderController extends Controller
      */
     public function index(Request $request)
     {
+        $design = BrandDesign::where('slug',$request->route('design'))->first();
+        if(!$design){
+            return abort(404);
+        }
         $my_brand = $request->session()->get('selected_brand', 'default');
-       $brand = Brand::with('slider_section')->where('slug',$my_brand->slug)->first();
-        return view('admin.Slider_home.listing',['slider' => $brand->slider_section, 'slug' => $brand->slug ]);
+        $slider = SliderSection::where(['brand_id' => $my_brand->id,'design_id' => $design->id])->get();
+        return view('admin.Slider_home.listing',['slider' => $slider, 'slug' => $my_brand->slug,'design' => $request->route('design') ]);
     }
     /**
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create(Request $request)
     {
+        $design = BrandDesign::where('slug',$request->route('design'))->first();
+        if(!$design){
+            return abort(404);
+        }
         $brand = new Brand();
-        return response()->view('admin/Slider_home.form');
+        return response()->view('admin/Slider_home.form',['design' => $request->route('design')]);
     }
     /**
      * Store a newly created resource in storage.
@@ -44,6 +53,10 @@ class MyBrandSliderController extends Controller
      */
     public function store(Request $request)
     {
+        $design = BrandDesign::where('slug',$request->route('design'))->first();
+        if(!$design){
+            return abort(404);
+        }
         if(!$request->hasFile('file_path')) {
         return Redirect::back()->withErrors([
                 'msg' => 'Image or video file is not attached'
@@ -72,7 +85,8 @@ class MyBrandSliderController extends Controller
         $brand = $request->session()->get('selected_brand', 'default');
 
         $slider = new SliderSection();
-        $request->merge(['brand_id' => $brand->id]);
+        $design = BrandDesign::where('slug',$request->route('design'))->first();
+        $request->merge(['brand_id' => $brand->id,'design_id' => $design->id ]);
         $slider->fill($request->input());
         $slider->save();
         $slider->file_path = Storage::disk('public')->putFile('slider_section', $request->file('file_path'));
@@ -80,17 +94,21 @@ class MyBrandSliderController extends Controller
         $slider->save();
         $my_brand = $request->session()->get('selected_brand', 'default');
         return redirect()
-            ->route('admin.mybrand.slider',['slug' => $my_brand->slug])
+            ->route('admin.mybrand.slider',['slug' => $my_brand->slug,'design' => $request->route('design')])
             ->with('success', 'New slide has been created');
     }
     public function destroy(Request $request){
 
+        $design = BrandDesign::where('slug',$request->route('design'))->first();
+        if(!$design){
+            return abort(404);
+        }
         $slider = SliderSection::find($request->input('slider'));
         Storage::disk('public')->delete($slider->file_path);
         $slider->delete();
         $brand = $request->session()->get('selected_brand');
         return redirect()
-            ->route('admin.mybrand.slider',['slug'=> $brand->slug ])
+            ->route('admin.mybrand.slider',['slug'=> $brand->slug ,'design' => $request->route('design')])
             ->with('success', 'Slide has been deleted');
     }
 }
